requirements.txt
pandas
scikit-learn
pyyaml
matplotlib
seaborn
joblib
sentence-transformers
faiss-cpu
ollama
requests
xgboost
lightgbm
langchain
langchain-community

